emmm

+++ scripts/bash/install-opencv.sh
@@ -5,8 +5,9 @@
 # -------------------------------------------------------------------- |
 #                       SCRIPT OPTIONS                                 |
 # ---------------------------------------------------------------------|
-OPENCV_VERSION='4.5.4'       # Version to be installed
-OPENCV_CONTRIB='YES'          # Install OpenCV's extra modules (YES/NO)
+OPENCV_VERSION='4.5.5' # Version to be installed
+OPENCV_CONTRIB='YES'   # Install OpenCV's extra modules (YES/NO)
+WITH_GTK='YES'
 # -------------------------------------------------------------------- |
 
 # |          THIS SCRIPT IS TESTED CORRECTLY ON          |
@@ -27,15 +28,12 @@ OPENCV_CONTRIB='YES'          # Install OpenCV's extra modules (YES/NO)
 # | Ubuntu 18.04 LTS | OpenCV 3.4.2 | OK   | 18 Jul 2018 |
 # | Debian 9.5       | OpenCV 3.4.2 | OK   | 18 Jul 2018 |
 
-
-
 # 1. KEEP UBUNTU OR DEBIAN UP TO DATE
 
 sudo apt-get -y update
-# sudo apt-get -y upgrade       # Uncomment to install new versions of packages currently installed
-# sudo apt-get -y dist-upgrade  # Uncomment to handle changing dependencies with new vers. of pack.
-# sudo apt-get -y autoremove    # Uncomment to remove packages that are now no longer needed
-
+sudo apt-get -y upgrade      # Uncomment to install new versions of packages currently installed
+sudo apt-get -y dist-upgrade # Uncomment to handle changing dependencies with new vers. of pack.
+sudo apt-get -y autoremove   # Uncomment to remove packages that are now no longer needed
 
 # 2. INSTALL THE DEPENDENCIES
 
@@ -43,23 +41,27 @@ sudo apt-get -y update
 sudo apt-get install -y build-essential cmake
 
 # GUI (if you want GTK, change 'qt5-default' to 'libgtkglext1-dev' and remove '-DWITH_QT=ON'):
-sudo apt-get install -y qt5-default libvtk6-dev
+if [ ${WITH_GTK} = 'YES' ]; then
+    sudo apt install -y libgtkglext1-dev libvtk6-dev
+else
+    sudo apt-get install -y qt5-default libvtk6-dev
+fi
 
 # Media I/O:
 sudo apt-get install -y zlib1g-dev libjpeg-dev libwebp-dev libpng-dev libtiff5-dev libjasper-dev \
-                        libopenexr-dev libgdal-dev
+    libopenexr-dev libgdal-dev
 
 # Video I/O:
 sudo apt-get install -y libdc1394-22-dev libavcodec-dev libavformat-dev libswscale-dev \
-                        libtheora-dev libvorbis-dev libxvidcore-dev libx264-dev yasm \
-                        libopencore-amrnb-dev libopencore-amrwb-dev libv4l-dev libxine2-dev
+    libtheora-dev libvorbis-dev libxvidcore-dev libx264-dev yasm \
+    libopencore-amrnb-dev libopencore-amrwb-dev libv4l-dev libxine2-dev
 
 # Parallelism and linear algebra libraries:
 sudo apt-get install -y libtbb-dev libeigen3-dev
 
 # Python:
-sudo apt-get install -y python-dev  python-tk  pylint  python-numpy  \
-                        python3-dev python3-tk pylint3 python3-numpy flake8
+sudo apt-get install -y python2-dev python2-tk pylint2 python2-numpy \
+    python3-dev python3-tk pylint3 python3-numpy flake8
 
 # Java:
 sudo apt-get install -y ant default-jdk
@@ -69,35 +71,82 @@ sudo apt-get install -y doxygen unzip wget
 
 #;s
 # 3. INSTALL THE LIBRARY
+if [ ! -d "OpenCV" ]; then
+    if [ ! -f "opencv-${OPENCV_VERSION}.zip"]; then
+        wget https://github.com/opencv/opencv/archive/${OPENCV_VERSION}.zip
+    else
+        mv opencv-${OPENCV_VERSION}.zip ${OPENCV_VERSION}.zip
+    fi
 
-# wget https://github.com/opencv/opencv/archive/${OPENCV_VERSION}.zip
-# unzip ${OPENCV_VERSION}.zip && rm ${OPENCV_VERSION}.zip
-# mv opencv-${OPENCV_VERSION} OpenCV
+    unzip ${OPENCV_VERSION}.zip && rm ${OPENCV_VERSION}.zip
+    mv opencv-${OPENCV_VERSION} OpenCV
 
-if [ $OPENCV_CONTRIB = 'YES' ]; then
-  wget https://github.com/opencv/opencv_contrib/archive/${OPENCV_VERSION}.zip
-  unzip ${OPENCV_VERSION}.zip && rm ${OPENCV_VERSION}.zip
-  mv opencv_contrib-${OPENCV_VERSION} opencv_contrib
-  mv opencv_contrib OpenCV
-fi
+else
 
-cd OpenCV && mkdir build && cd build
+    echo "dir 'OpenCV' is existing"
 
-if [ $OPENCV_CONTRIB = 'NO' ]; then
-cmake -DWITH_QT=ON -DWITH_OPENGL=ON -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
-      -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF ..
 fi
 
 if [ $OPENCV_CONTRIB = 'YES' ]; then
-cmake -DWITH_QT=ON -DWITH_OPENGL=ON -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
-      -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF \
-      -DOPENCV_EXTRA_MODULES_PATH=../opencv_contrib/modules ..
+    if [ ! -d "OpenCV/opencv_contrib" ]; then
+        if [ ! -f "opencv_contrib-${OPENCV_VERSION}.zip"]; then
+            wget https://github.com/opencv/opencv_contrib/archive/${OPENCV_VERSION}.zip
+        else
+            mv opencv_contrib-${OPENCV_VERSION}.zip ${OPENCV_VERSION}.zip
+        fi
+        unzip ${OPENCV_VERSION}.zip && rm ${OPENCV_VERSION}.zip
+        mv opencv_contrib-${OPENCV_VERSION} opencv_contrib
+        mv opencv_contrib OpenCV
+    else
+        echo "dir 'opencv_contrib' is existing"
+    fi
+fi
+#------------------------------------------------------------------------------
+
+cd OpenCV
+if [ ! -d build ]; then
+    mkdir build && cd build
+else
+    cd build
 fi
 
-make -j3
+if [ ${WITH_GTK}='YES' ]; then
+
+    if [ $OPENCV_CONTRIB = 'NO' ]; then
+        cmake -DWITH_OPENGL=ON -DWITH_GTK:BOOL=ON -DWITH_GTK_2_X:BOOL=ON \
+            -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
+            -DBUILD_examples=OFF \
+            -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF ..
+    fi
+
+    if [ $OPENCV_CONTRIB = 'YES' ]; then
+        cmake -DWITH_OPENGL=ON -DWITH_GTK:BOOL=ON -DWITH_GTK_2_X:BOOL=ON \
+        -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
+        -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF \
+        -DBUILD_examples=OFF -DOpenGL_GL_PREFERENCE=GLVND\
+        -DOPENCV_EXTRA_MODULES_PATH=../opencv_contrib/modules ..
+    fi
+
+else #...............................................
+
+    if [ $OPENCV_CONTRIB = 'NO' ]; then
+        cmake -DWITH_QT=ON -DWITH_OPENGL=ON -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
+            -DBUILD_examples=OFF \
+            -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF ..
+    fi
+
+    if [ $OPENCV_CONTRIB = 'YES' ]; then
+        cmake -DWITH_QT=ON -DWITH_OPENGL=ON -DFORCE_VTK=ON -DWITH_TBB=ON -DWITH_GDAL=ON \
+            -DWITH_XINE=ON -DENABLE_PRECOMPILED_HEADERS=OFF \
+            -DBUILD_examples=OFF -DOpenGL_GL_PREFERENCE=GLVND\
+            -DOPENCV_EXTRA_MODULES_PATH=../opencv_contrib/modules ..
+    fi
+fi
+
+#------------------------------------------------------------------------------
+make -j6
 sudo make install
 sudo ldconfig
-
 
 # 4. EXECUTE SOME OPENCV EXAMPLES AND COMPILE A DEMONSTRATION
